requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-HTTPAuth==4.3.0
PyJWT==2.3.0
Werkzeug==2.1.1
